requirements.txt
asgiref==3.4.1
auth==0.5.3
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-material==1.10.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
eventlet==0.33.0
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
material==0.1
mongoengine==0.23.1
myapi==0.0.1
openapi-codec==1.3.2
Pillow==9.0.0
psycopg2==2.9.3
Pygments==2.10.0
PyJWT==2.3.0
pymongo==3.12.3
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
